* Refactoring for readability * Move lookup() code to connectors
